First commit
The AI-Based Disease Prediction System is a comprehensive tool designed to enhance healthcare by providing accurate and timely disease predictions. By leveraging the power of machine learning and the robustness of the Django framework and PostgreSQL, this system aims to improve patient outcomes and streamline the diagnostic process for healthcare providers.